feat(evasive-transform): expose evadeCensorSync
This change replaces [source-map](https://npm.im/source-map) with [source-map-js](https://npm.im/source-map-js), which is a fork of the former.  Crucially, `source-map-js` is a synchronous, pure-JS implementation.

A consequence of this is that `makeLocationUnmapper` is now synchronous.  It is internal, however.

`evadeCensor` now just wraps `evadeCensorSync` in a `Promise`.  Tests have been changed to use `evadeCensorSync` directly.

In the `makeLocationUnmapper` implementation, an assertion for the truthiness of `ast.loc` has been moved _before_ instantiation of `SourceMapConsumer`, where it maybe should have been in the first place.